6575640 rename/rmdir/remove on mixed file systems work incorrectly for mangled names 6582182 64-bit libraries 6588720 Unable to create new folder in autohome share from Solaris10u3 with UFS & ZFS. 6611820 Unable to map a SMB share on SPARC system. 6612193 Rework ADS client, DynDNS client and Auth IPC to use host principal 6614664 CIFS server should not request ctime changes 6614864 Security tab not available after a share is mapped on SPARC system 6614953 An error occurred while creating a notepad file on X86 32-bit system. 6616349 Sparc-Get error "input not UTF-8" and (Abort)(core dumped) in sharemgr with Korean resource name 6622828 Resolve unreferenced makefiles 6626833 SUNWsmbsu package issues with snv_77. 6627415 SmbCreateTemporary fails on SPARC 6627469 Unable to map SIDs to names for Windows Vista clients 6627471 Remove winpipe door server from name space 6628747 smbd crash and in maintenance state 6629402 The adjoin operation triggered by smbadm join CLI shouldn't overwrite the Kerberos keytab file. 6629410 SMB daemon should remove its ccache on refresh or restart 6629428 smb daemon crashes at smb_nic_clear_niclist+0x45(815ff10, 2) 6629797 sharemgr: Deleting an SMB resource that has an option leads to core dump 6630470 Get core in sharemgr:libshare_smb.so.1`smb_enable_service+0x78(2ec90, fe87c49c, 0, ff396000,17e54,6)
--HG--
rename : usr/src/uts/common/smbsrv/smb_secdesc.h => deleted_files/usr/src/uts/common/smbsrv/smb_secdesc.h

+++ usr/src/lib/smbsrv/libmlsvc/common/mlsvc_srvsvc.c
@@ -512,7 +512,8 @@ srvsvc_s_NetShareSetInfo(void *arg, struct mlrpc_xaction *mxa)
 	struct mlsm_NetShareSetInfo *param = arg;
 
 	(void) memset(param, 0, sizeof (struct mlsm_NetShareSetInfo));
-	param->parm_err_ptr = (DWORD)MLRPC_HEAP_MALLOC(mxa, sizeof (DWORD));
+	param->parm_err_ptr = (DWORD)(uintptr_t)MLRPC_HEAP_MALLOC(mxa,
+	    sizeof (DWORD));
 	param->parm_err = 0;
 
 	smb_config_rdlock();
